requirements.txt
click==8.1.3
Flask==2.2.2
greenlet==3.0.3
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==1.0.1
SQLAlchemy==1.4.43
typing_extensions==4.4.0
Werkzeug==2.2.2
zipp==3.10.0
